+++ src/python/autotransform/util/scheduler.py
@@ -13,7 +13,6 @@ from __future__ import annotations
 
 import json
 import os
-from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
 from typing import Any, Dict, List, Optional
@@ -30,14 +29,16 @@ from autotransform.runner.base import Runner
 from autotransform.runner.github import GithubRunner
 from autotransform.schema.builder import FACTORY as schema_builder_factory
 from autotransform.schema.schema import AutoTransformSchema
+from autotransform.util.component import ComponentModel
 from autotransform.util.console import (
     choose_option,
+    choose_options_from_list,
     choose_yes_or_no,
     error,
     get_str,
     info,
     input_int,
-    input_string,
+    input_ints,
 )
 
 
@@ -55,8 +56,7 @@ class RepeatSetting(str, Enum):
     WEEKLY = "weekly"
 
 
-@dataclass(kw_only=True)
-class SchemaScheduleSettings:
+class SchemaScheduleSettings(ComponentModel):
     """A schedule for a Schema.
 
     Attributes:
@@ -64,14 +64,14 @@ class SchemaScheduleSettings:
         hour_of_day (int): The hour of the day to run the Schema, 0-23.
         day_of_week (Optional[int], optional): The day of the week to run a weekly Schema.
             Defaults to None.
-        sharding (Optional[ShardingSettings], optional): The settings used to determine how
-            to apply sharding to Schema runs. Defaults to None.
+        shard_filter (Optional[ShardingSettings], optional): A shard filter to use for sharding
+            the inputs across multiple runs. Defaults to None.
     """
 
     repeats: RepeatSetting
     hour_of_day: int
     day_of_week: Optional[int] = None
-    sharding: Optional[ShardFilter] = None
+    shard_filter: Optional[ShardFilter] = None
 
     def should_run(self, hour_of_day: int, day_of_week: int) -> bool:
         """Whether the schema should be run.
@@ -88,35 +88,12 @@ class SchemaScheduleSettings:
             self.repeats == RepeatSetting.DAILY or self.day_of_week == day_of_week
         )
 
-    def bundle(self) -> Dict[str, Any]:
-        """Generates a JSON encodable bundle.
-
-        Returns:
-            Dict[str, Any]: The encodable bundle.
-        """
-
-        bundle: Dict[str, Any] = {
-            "repeats": self.repeats.value,
-            "hour_of_day": self.hour_of_day,
-        }
-
-        if self.day_of_week is not None:
-            bundle["day_of_week"] = self.day_of_week
-
-        if self.sharding is not None:
-            shard_bundle = self.sharding.bundle()
-            del shard_bundle["valid_shard"]
-            bundle["sharding"] = shard_bundle
-
-        return bundle
-
     @staticmethod
-    def from_data(data: Dict[str, Any], elapsed_days: int) -> SchemaScheduleSettings:
+    def from_data(data: Dict[str, Any]) -> SchemaScheduleSettings:
         """Produces an instance of the SchemaScheduleSettings from decoded data.
 
         Args:
             data (Mapping[str, Any]): The JSON decoded data.
-            elapsed_days (int): The number of days elapsed since the start of the schedule.
 
         Returns:
             SchemaScheduleSettings: An instance of the SchemaScheduleSettings.
@@ -131,24 +108,17 @@ class SchemaScheduleSettings:
             assert isinstance(day_of_week, int)
             assert day_of_week in range(7)
 
-        sharding = data.get("sharding", None)
-        if sharding is not None:
-            assert isinstance(sharding, Dict)
-            num_shards = sharding["num_shards"]
-            assert isinstance(num_shards, int)
-            if repeats == RepeatSetting.DAILY:
-                valid_shard = elapsed_days % num_shards
-            else:
-                valid_shard = (elapsed_days // 7) % num_shards
-            sharding["valid_shard"] = valid_shard
-            filter_instance = filter_factory.get_instance(sharding)
-            assert isinstance(filter_instance, ShardFilter)
-            shard_filter = filter_instance
+        shard_filter = data.get("shard_filter", None)
+        if shard_filter is not None:
+            shard_filter = filter_factory.get_instance(shard_filter)
         else:
             shard_filter = None
 
         return SchemaScheduleSettings(
-            repeats=repeats, hour_of_day=hour_of_day, day_of_week=day_of_week, sharding=shard_filter
+            repeats=repeats,
+            hour_of_day=hour_of_day,
+            day_of_week=day_of_week,
+            shard_filter=shard_filter,
         )
 
     @staticmethod
@@ -181,76 +151,60 @@ class SchemaScheduleSettings:
         else:
             day_of_week = None
 
-        # Get sharding
+        # Get shard_filter
         if not simple and choose_yes_or_no("Do you want to use sharding for this schema?"):
-            default_filter = KeyHashShardFilter(num_shards=10, valid_shard=0).bundle()
-            del default_filter["valid_shard"]
-            shard_filter: Optional[ShardFilter] = None
-            while shard_filter is None:
-                shard_filter_json = input_string(
-                    "Enter the JSON encoded shard filter:",
+            shard_filter = filter_factory.from_console(
+                "shard filter",
+                default_value=KeyHashShardFilter(num_shards=10),
+                allow_none=False,
+            )
+            while not isinstance(shard_filter, ShardFilter):
+                assert shard_filter is not None
+                error(f"{shard_filter.name} is not a shard filter")
+                shard_filter = filter_factory.from_console(
                     "shard filter",
-                    default=json.dumps(default_filter),
+                    allow_none=False,
                 )
-                try:
-                    shard_filter_bundle = json.loads(shard_filter_json)
-                    if "valid_shard" not in shard_filter_bundle:
-                        shard_filter_bundle["valid_shard"] = 0
-                    filter_instance = filter_factory.get_instance(shard_filter_bundle)
-                    if isinstance(filter_instance, ShardFilter):
-                        shard_filter = filter_instance
-                    else:
-                        error("Invalid filter provided, please provide a shard filter.")
-                except Exception:  # pylint: disable=broad-except
-                    error("Failed to get shard filter from input, please try again.")
         else:
             shard_filter = None
 
         return SchemaScheduleSettings(
-            repeats=repeats, hour_of_day=hour_of_day, day_of_week=day_of_week, sharding=shard_filter
+            repeats=repeats,
+            hour_of_day=hour_of_day,
+            day_of_week=day_of_week,
+            shard_filter=shard_filter,
         )
 
 
-@dataclass(kw_only=True)
-class ScheduledSchema:
+class ScheduledSchema(ComponentModel):
     """A Schema that is scheduled for automatic runs.
 
     Attributes:
         type (SchemaType): The type of Schema that is specified, file or builder.
-        schema (str): The Schema that is being scheduled.
+        target (str): The Schema that is being scheduled.
         schedule (SchemaScheduleSettings): The settings used to determine when to run the Schema.
     """
 
     type: SchemaType
-    schema: str
+    target: str
     schedule: SchemaScheduleSettings
 
-    def bundle(self) -> Dict[str, Any]:
-        """Generates a JSON encodable bundle.
-
-        Returns:
-            Dict[str, Any]: The encodable bundle.
-        """
-
-        return {"type": self.type.value, "schema": self.schema, "schedule": self.schedule.bundle()}
-
     @staticmethod
-    def from_data(data: Dict[str, Any], elapsed_days: int) -> ScheduledSchema:
+    def from_data(data: Dict[str, Any]) -> ScheduledSchema:
         """Produces an instance of the ScheduledSchema from decoded data.
 
         Args:
             data (Mapping[str, Any]): The JSON decoded data.
-            elapsed_days (int): The number of days elapsed since the start of the schedule.
 
         Returns:
             ScheduledSchema: An instance of the ScheduledSchema.
         """
 
         schema_type = SchemaType(data["type"])
-        schema = data["schema"]
-        assert isinstance(schema, str)
-        schedule = SchemaScheduleSettings.from_data(data["schedule"], elapsed_days)
-        return ScheduledSchema(type=schema_type, schema=schema, schedule=schedule)
+        target = data["target"]
+        assert isinstance(target, str)
+        schedule = SchemaScheduleSettings.from_data(data["schedule"])
+        return ScheduledSchema(type=schema_type, target=target, schedule=schedule)
 
     @staticmethod
     def from_console() -> ScheduledSchema:
@@ -260,7 +214,7 @@ class ScheduledSchema:
             ScheduledSchema: The input ScheduledSchema.
         """
 
-        schema = get_str("Enter the schema to schedule: ")
+        target = get_str("Enter the schema to schedule: ")
 
         schema_type = choose_option(
             "What is the type of the schema?",
@@ -268,12 +222,13 @@ class ScheduledSchema:
         )
 
         return ScheduledSchema(
-            schema=schema, type=schema_type, schedule=SchemaScheduleSettings.from_console()
+            target=target,
+            type=schema_type,
+            schedule=SchemaScheduleSettings.from_console(),
         )
 
 
-@dataclass(kw_only=True)
-class Scheduler:
+class Scheduler(ComponentModel):
     """The information and functionality required to schedule Schemas.
 
     Attributes:
@@ -323,14 +278,14 @@ class Scheduler:
             # Get the Schema
             if scheduled_schema.type == SchemaType.BUILDER:
                 try:
-                    schema = json.loads(scheduled_schema.schema)
+                    schema = json.loads(scheduled_schema.target)
                 except json.JSONDecodeError:
-                    schema = {"name": scheduled_schema.schema}
+                    schema = {"name": scheduled_schema.target}
                 if isinstance(schema, str):
                     schema = {"name": schema}
                 schema = schema_builder_factory.get_instance(schema).build()
             else:
-                with open(scheduled_schema.schema, "r", encoding="UTF-8") as schema_file:
+                with open(scheduled_schema.target, "r", encoding="UTF-8") as schema_file:
                     schema = AutoTransformSchema.from_json(schema_file.read())
 
             # Check if should run
@@ -341,8 +296,12 @@ class Scheduler:
                     )
                 )
                 continue
-            shard_filter = scheduled_schema.schedule.sharding
+            shard_filter = scheduled_schema.schedule.shard_filter
             if shard_filter is not None:
+                if scheduled_schema.schedule.repeats == RepeatSetting.DAILY:
+                    shard_filter.valid_shard = elapsed_days % shard_filter.num_shards
+                else:
+                    shard_filter.valid_shard = (elapsed_days // 7) % shard_filter.num_shards
                 EventHandler.get().handle(
                     DebugEvent(
                         {
@@ -369,27 +328,12 @@ class Scheduler:
             scheduler_file.write(json.dumps(self.bundle(), indent=4))
             scheduler_file.flush()
 
-    def bundle(self) -> Dict[str, Any]:
-        """Generates a JSON encodable bundle.
-
-        Returns:
-            Dict[str, Any]: The encodable bundle.
-        """
-
-        return {
-            "base_time": self.base_time,
-            "excluded_days": self.excluded_days,
-            "runner": self.runner.bundle(),
-            "schemas": [schema.bundle() for schema in self.schemas],
-        }
-
     @staticmethod
-    def read(file_path: str, start_time: int) -> Scheduler:
+    def read(file_path: str) -> Scheduler:
         """Reads a Scheduler from a JSON encoded file.
 
         Args:
             file_path (str): The path where the JSON for the Schedule is located.
-            start_time (int): The start time to use for setting up sharding/running.
 
         Returns:
             Scheduler: The Scheduler from the file.
@@ -400,10 +344,10 @@ class Scheduler:
         EventHandler.get().handle(
             DebugEvent({"message": f"Scheduler: ({file_path})\n{scheduler_json}"})
         )
-        return Scheduler.from_json(scheduler_json, start_time)
+        return Scheduler.from_json(scheduler_json)
 
     @staticmethod
-    def from_json(scheduler_json: str, start_time: int) -> Scheduler:
+    def from_json(scheduler_json: str) -> Scheduler:
         """Builds a Scheduler from JSON encoded values.
 
         Args:
@@ -414,15 +358,14 @@ class Scheduler:
             Scheduler: The Scheduler from the JSON.
         """
 
-        return Scheduler.from_data(json.loads(scheduler_json), start_time)
+        return Scheduler.from_data(json.loads(scheduler_json))
 
     @staticmethod
-    def from_data(data: Dict[str, Any], start_time: int) -> Scheduler:
+    def from_data(data: Dict[str, Any]) -> Scheduler:
         """Produces an instance of the Scheduler from decoded data.
 
         Args:
             data (Mapping[str, Any]): The JSON decoded data.
-            start_time (int): The time that the schedule is starting to run on.
 
         Returns:
             Scheduler: An instance of the Scheduler.
@@ -436,16 +379,13 @@ class Scheduler:
             assert isinstance(day, int)
             assert day in range(7)
         runner = runner_factory.get_instance(data["runner"])
-        elapsed_days = (start_time - base_time) // 60 // 60 // 24
-        schemas = [
-            ScheduledSchema.from_data(schema, elapsed_days) for schema in data.get("schemas", [])
-        ]
+        schemas = [ScheduledSchema.from_data(schema) for schema in data.get("schemas", [])]
         return Scheduler(
             base_time=base_time, excluded_days=excluded_days, runner=runner, schemas=schemas
         )
 
     @staticmethod
-    def from_console(
+    def init_from_console(
         runner: Optional[Runner] = None, use_sample_schema: bool = False, simple: bool = False
     ) -> Scheduler:
         """Gets a Scheduler using console input.
@@ -475,7 +415,7 @@ class Scheduler:
             excluded_days = [5, 6]
 
         # Gets the Runner
-        input_runner = runner_factory.from_console(
+        runner = runner_factory.from_console(
             "scheduler runner",
             previous_value=runner,
             default_value=GithubRunner(
@@ -485,7 +425,7 @@ class Scheduler:
             simple=simple,
             allow_none=False,
         )
-        assert input_runner is not None
+        assert runner is not None
 
         # Gets Schemas
         if use_sample_schema:
@@ -493,7 +433,7 @@ class Scheduler:
             schemas = [
                 ScheduledSchema(
                     type=SchemaType.FILE,
-                    schema=f"{relative_path}/schemas/black_format.json",
+                    target=f"{relative_path}/schemas/black_format.json",
                     schedule=SchemaScheduleSettings(repeats=RepeatSetting.DAILY, hour_of_day=7),
                 )
             ]
@@ -507,6 +447,70 @@ class Scheduler:
             base_time=int(datetime.fromisoformat("2022-05-23T00:00:00").timestamp())
             + base_modifier,
             excluded_days=excluded_days,
-            runner=input_runner,
+            runner=runner,
             schemas=schemas,
+        )
+
+    @staticmethod
+    def from_console(prev_scheduler: Optional[Scheduler] = None) -> Scheduler:
+        """Gets a Scheduler using console inputs.
+
+        Args:
+            prev_scheduler (Optional[Scheduler], optional): A previously input Scheduler.
+                Defaults to None.
+
+        Returns:
+            Scheduler: The input Scheduler.
+        """
+
+        # Get Base Time
+        if prev_scheduler and choose_yes_or_no(
+            f"Use previous base time ({prev_scheduler.base_time})?"
+        ):
+            base_time = prev_scheduler.base_time
+        else:
+            base_time = int(datetime.fromisoformat("2022-05-23T00:00:00").timestamp())
+            info("Using local time to establish a base")
+            info("Midnight Monday local is day_of_week 0, hour_of_day 0")
+            info(f"Associated timestamp is {base_time}")
+            if choose_yes_or_no("Apply a modifier to local time for scheduling?"):
+                base_time += input_int("Enter the modifier in seconds")
+
+        # Get Runner
+        args: Dict[str, Any] = {
+            "default_value": GithubRunner(
+                run_workflow="autotransform.run.yml",
+                update_workflow="autotransform.update.yml",
+            ),
+            "allow_none": False,
+        }
+        if prev_scheduler is not None:
+            args["previous_value"] = prev_scheduler.runner
+        runner = runner_factory.from_console("scheduler runner", **args)
+        assert runner is not None
+
+        # Get Excluded Days
+        if prev_scheduler is not None and choose_yes_or_no(
+            f"Use previous excluded days: {prev_scheduler.excluded_days!r}"
+        ):
+            excluded_days = prev_scheduler.excluded_days
+        else:
+            excluded_days = input_ints(
+                "Enter excluded days", min_val=0, max_val=6, min_choices=0, max_choices=7
+            )
+
+        # Get Schemas
+        if prev_scheduler is not None and bool(prev_scheduler.schemas):
+            schemas = choose_options_from_list(
+                "Choose schemas to keep",
+                [(schema, f"{schema!r}") for schema in prev_scheduler.schemas],
+                min_choices=0,
+                max_choices=len(prev_scheduler.schemas),
+            )
+        else:
+            schemas = []
+        while choose_yes_or_no("Add a schema to the schedule?"):
+            schemas.append(ScheduledSchema.from_console())
+        return Scheduler(
+            base_time=base_time, excluded_days=excluded_days, runner=runner, schemas=schemas
         )